Trunk 1.1: zipios fixes and updates - 64 bit support fix - merge of "compatible" zipios cvs trunk diffs up to 3 Sep 07 Note: - zipios API change on zipinputstream is not compatible   (src/zipfile.cpp, src/zipinputstreambuf.cpp,   zipios++/zipinputstream.h diffs excluded) - Enigma zipios version included additional OS support and   compatibility fixes

+++ lib-src/zipios++/zipios++/zipios_defs.h
@@ -1,10 +1,13 @@
 #ifndef ZIPIOS_DEFS_H
 #define ZIPIOS_DEFS_H
 
+#include "zipios++/zipios-config.h"
+#include <stdint.h>
+
 namespace zipios {
 
-typedef unsigned short uint16 ;
-typedef unsigned long  uint32 ;
+typedef uint16_t uint16 ;
+typedef uint32_t uint32 ;
 
 } // namespace
